+++ app/config/lifespan.py
@@ -1,0 +1,9 @@
+from fastapi import FastAPI
+from contextlib import asynccontextmanager
+from settings.db import ensure_db_connection
+
+
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    await ensure_db_connection()
+    yield
