+++ requirements.txt
@@ -2,13 +2,16 @@ annotated-doc==0.0.3
 annotated-types==0.7.0
 anyio==4.11.0
 click==8.1.8
+dnspython==2.8.0
 exceptiongroup==1.3.0
 fastapi==0.120.2
 h11==0.16.0
 httptools==0.7.1
 idna==3.11
+motor==3.7.1
 pydantic==2.12.3
 pydantic_core==2.41.4
+pymongo==4.15.3
 python-dotenv==1.2.1
 PyYAML==6.0.3
 sniffio==1.3.1
